modify agent prompt; add python_repl tool; adapt to pydantic v2
