+++ source/lambda/online/common_logic/common_utils/lambda_invoke_utils.py
@@ -3,18 +3,21 @@ import functools
 import importlib
 import json
 import time
+import os 
 from typing import Any, Dict, Optional, Callable, Union
+import threading
 
 import requests
 from common_logic.common_utils.constant import StreamMessageType
 from common_logic.common_utils.logger_utils import get_logger
 from common_logic.common_utils.websocket_utils import is_websocket_request, send_to_ws_client
-from langchain.pydantic_v1 import BaseModel, Field, root_validator
+from pydantic import BaseModel, Field, model_validator
+
 
 from .exceptions import LambdaInvokeError
 
 logger = get_logger("lambda_invoke_utils")
-
+thread_local = threading.local()
 
 __FUNC_NAME_MAP = {
     "query_preprocess": "Preprocess for multi-round conversation",
@@ -24,6 +27,33 @@ __FUNC_NAME_MAP = {
     "results_evaluation": "Result evaluation",
     "tool_execution": "Final tool result"
 }
+
+
+class StateContext:
+
+    def __init__(self,state):
+        self.state=state
+    
+    @classmethod
+    def get_current_state(cls):
+        state = getattr(thread_local,'state',None)
+        assert state is not None,"There is not a valid state in current context"
+        return state
+
+    @classmethod
+    def set_current_state(cls, state):
+        setattr(thread_local, 'state', state)
+    
+    @classmethod
+    def clear_state(cls):
+        setattr(thread_local, 'state', None)
+
+    def __enter__(self):
+        self.set_current_state(self.state)
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.clear_state()
+
 
 class LAMBDA_INVOKE_MODE(enum.Enum):
     LAMBDA = "lambda"
@@ -54,26 +84,24 @@ class LambdaInvoker(BaseModel):
     region_name: str = None
     credentials_profile_name: Optional[str] = Field(default=None, exclude=True)
 
-    @root_validator()
+    @model_validator(mode="before")
     def validate_environment(cls, values: Dict):
         if values.get("client") is not None:
             return values
         try:
             import boto3
-
             try:
-                if values["credentials_profile_name"] is not None:
+                if values.get("credentials_profile_name") is not None:
                     session = boto3.Session(
                         profile_name=values["credentials_profile_name"]
                     )
                 else:
                     # use default credentials
                     session = boto3.Session()
-
                 values["client"] = session.client(
-                    "lambda", region_name=values["region_name"]
+                    "lambda",
+                    region_name=values.get("region_name",os.environ['AWS_REGION'])
                 )
-
             except Exception as e:
                 raise ValueError(
                     "Could not load credentials to authenticate with AWS client. "
@@ -284,7 +312,10 @@ def node_monitor_wrapper(fn: Optional[Callable[..., Any]] = None, *, monitor_key
                        current_stream_use, ws_connection_id, enable_trace)
             state['trace_infos'].append(
                 f"Enter: {func.__name__}, time: {time.time()}")
-            output = func(state)
+            
+            with StateContext(state): 
+                output = func(state)
+
             current_monitor_infos = output.get(monitor_key, None)
             if current_monitor_infos is not None:
                 send_trace(f"\n\n {current_monitor_infos}",
