+++ source/lambda/online/common_logic/langchain_integration/tools/__init__.py
@@ -1,33 +1,3 @@
-# from langchain.tools.base import StructuredTool,BaseTool,tool
-# StructuredTool.from_function
-# from langchain_experimental.tools import PythonREPLTool
-# from langchain_core.utils.function_calling import convert_to_openai_function
-# from llama_index.core.tools import FunctionTool
-# from langchain.tools import BaseTool
-# from pydantic import create_model
-
-# from langchain_community.tools import WikipediaQueryRun
-
-
-# builder = StateGraph(State)
-
-
-# # Define nodes: these do the work
-# builder.add_node("assistant", Assistant(part_1_assistant_runnable))
-# builder.add_node("tools", create_tool_node_with_fallback(part_1_tools))
-# # Define edges: these determine how the control flow moves
-# builder.add_edge(START, "assistant")
-# builder.add_conditional_edges(
-#     "assistant",
-#     tools_condition,
-# )
-# builder.add_edge("tools", "assistant")
-
-# # The checkpointer lets the graph persist its state
-# # this is a complete memory for the entire graph.
-# memory = MemorySaver()
-# part_1_graph = builder.compile(checkpointer=memory)
-
 from typing import Optional,Union
 from pydantic import BaseModel
 import platform
@@ -40,7 +10,7 @@ from datamodel_code_generator import DataModelType, PythonVersion
 from datamodel_code_generator.model import get_data_model_types
 from datamodel_code_generator.parser.jsonschema import JsonSchemaParser
 from langchain.tools.base import StructuredTool as _StructuredTool ,BaseTool
-from langchain_core.pydantic_v1 import create_model,BaseModel
+# from langchain_core.pydantic_v1 import BaseModel
 from common_logic.common_utils.constant import SceneType
 from common_logic.common_utils.lambda_invoke_utils import invoke_with_lambda
 from functools import partial
@@ -48,10 +18,9 @@ from functools import partial
 
 
 class StructuredTool(_StructuredTool):
-    pass_state:bool = False # if pass state into tool invoke 
-    pass_state_name:str = "state" # pass state name 
-    
-
+    pass
+    # pass_state:bool = False # if pass state into tool invoke 
+    # pass_state_name:str = "state" # pass state name 
 
 class ToolIdentifier(BaseModel):
     scene: SceneType
@@ -89,7 +58,6 @@ class ToolManager:
         new_tool_module = types.ModuleType(tool_id)
         exec(result, new_tool_module.__dict__)
         model_cls = new_tool_module.Model
-        # model_cls.model_rebuild()
         return model_cls
 
     
@@ -191,8 +159,8 @@ class ToolManager:
         name=None,
         tool_identifier=None,   
         return_direct=False,
-        pass_state=True,
-        pass_state_name='state'
+        # pass_state=True,
+        # pass_state_name='state'
     ):
         assert scene == SceneType.COMMON, scene
         from .common_tools.rag import rag_tool
@@ -218,8 +186,8 @@ class ToolManager:
             ),
             description=description,
             return_direct=return_direct,
-            pass_state=pass_state,
-            pass_state_name=pass_state_name
+            # pass_state=pass_state,
+            # pass_state_name=pass_state_name
         )
         
         return ToolManager.register_lc_tool(
@@ -254,118 +222,7 @@ def lazy_tool_load_decorator(scene:SceneType,name):
     return decorator
 
 
-############################# tool load func ######################
-
-@lazy_tool_load_decorator(SceneType.COMMON,"get_weather")
-def _load_common_weather_tool(tool_identifier:ToolIdentifier):
-    from .common_tools import get_weather
-    tool_def = {
-        "description": "Get the current weather for `city_name`",
-        "properties": {
-            "city_name": {
-                "description": "The name of the city to be queried",
-                "type": "string"
-            },
-        },
-        "required": ["city_name"]
-    }
-    ToolManager.register_func_as_tool(
-        func=get_weather.get_weather,
-        tool_def=tool_def,
-        scene=tool_identifier.scene,
-        name=tool_identifier.name,
-        return_direct=False
-    )
-
-
-@lazy_tool_load_decorator(SceneType.COMMON,"give_rhetorical_question")
-def _load_common_rhetorical_tool(tool_identifier:ToolIdentifier):
-    from .common_tools import give_rhetorical_question
-    tool_def = {
-        "description": "If the user's question is not clear and specific, resulting in the inability to call other tools, please call this tool to ask the user a rhetorical question",
-        "properties": {
-            "question": {
-                "description": "The rhetorical question to user",
-                "type": "string"
-            },
-        },
-        "required": [] #["question"]
-    } 
-    ToolManager.register_func_as_tool(
-        scene=tool_identifier.scene,
-        name=tool_identifier.name,
-        func=give_rhetorical_question.give_rhetorical_question,
-        tool_def=tool_def,
-        return_direct=True
-    )
-
-
-@lazy_tool_load_decorator(SceneType.COMMON,"give_final_response")
-def _load_common_final_response_tool(tool_identifier:ToolIdentifier):
-    from .common_tools import give_final_response
-    
-    tool_def = {
-        "description": "If none of the other tools need to be called, call the current tool to complete the direct response to the user.",
-        "properties": {
-            "response": {
-                "description": "Response to user",
-                "type": "string"
-            }
-        },
-        "required": ["response"]
-    }
-    ToolManager.register_func_as_tool(
-        scene=tool_identifier.scene,
-        name=tool_identifier.name,
-        func=give_final_response.give_final_response,
-        tool_def=tool_def,
-        return_direct=True
-    )
-
-
-@lazy_tool_load_decorator(SceneType.COMMON,"chat")
-def _load_common_chat_tool(tool_identifier:ToolIdentifier):
-    from .common_tools import chat
-    tool_def = {
-        "description": "casual talk with AI",
-        "properties": {
-            "response": {
-                "description": "response to users",
-                "type": "string"
-                }
-        },
-        "required": ["response"]
-    }
-
-    ToolManager.register_func_as_tool(
-        scene=tool_identifier.scene,
-        name=tool_identifier.name,
-        func=chat.chat,
-        tool_def=tool_def,
-        return_direct=True
-    )
-
-
-@lazy_tool_load_decorator(SceneType.COMMON,"rag_tool")
-def _load_common_rag_tool(tool_identifier:ToolIdentifier):
-    from .common_tools import rag
-    tool_def = {
-        "description": "private knowledge",
-        "properties": {
-            "query": {
-                "description": "query for retrieve",
-                "type": "string"
-                }
-        },
-        # "required": ["query"]
-    }
-    ToolManager.register_func_as_tool(
-        scene=tool_identifier.scene,
-        name=tool_identifier.name,
-        func=rag.rag_tool,
-        tool_def=tool_def,
-        return_direct=True
-    )
+from . import common_tools
