+++ source/lambda/online/functions/__init__.py
@@ -1,11 +1,11 @@
 # tool
-from ._tool_base import get_tool_by_name,Tool,tool_manager
+# from ._tool_base import get_tool_by_name,Tool,tool_manager
 
-def init_common_tools():
-    from . import lambda_common_tools
+# def init_common_tools():
+#     from . import lambda_common_tools
 
-def init_aws_qa_tools():
-    from . import lambda_aws_qa_tools
+# def init_aws_qa_tools():
+#     from . import lambda_aws_qa_tools
 
-def init_retail_tools():
-    from . import lambda_retail_tools
+# def init_retail_tools():
+#     from . import lambda_retail_tools
